code/requirements.txt
dgl==2.3.0+cu121
jarvis_tools==2024.8.30
lmdb==1.5.1
networkx==3.3
numpy==2.2.2
pandas==2.2.3
rdkit==2024.3.5
rdkit_pypi==2022.9.5
scikit_learn==1.6.1
scipy==1.15.1
torch==2.3.1
torch_scatter==2.1.2+pt23cu121
tqdm==4.66.5